suggestion of minor change in the paradigm waiting message
Sometimes the sentence that is created by the code is incorrect. For example:
"Waiting for a RPC-FST server start." is incorrect.

It should be "Waiting for an RPC-FST server start."

Creating a method to select the proper "a" or "an" is very difficult to write, as for example r is a consonant and yet it is pronounced as a vowel in many cases such as this one. It would require a phonetic dictionary etc.

For this reason I propose to change this message to:
"Waiting for the RPC-FST server to start." thus avoiding "a" and "an".
minor fix